% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fun.R
\docType{data}
\name{base_digits}
\alias{base_digits}
\title{Base digits (as named vector)}
\format{
An object of class \code{character} of length 16.
}
\usage{
base_digits
}
\description{
\code{base_digits} provides numeral symbol for common bases 
as a named character vector.
}
\details{
Note that the elements are character symbols ("0"-"9", "A"-"F"), 
whereas the names coincide to numeric values (from 0 to 15).
}
\examples{
base_digits
length(base_digits)  # 16, but zero-indexed names
base_digits[10]      # 10. element ("9" with name "9") 
base_digits["10"]    # named element "10" ("A" with name "10")
base_digits[["10"]]  # element named "10" ("A")

}
\seealso{
\code{\link{base2dec}} converts numerals in some base into decimal numbers; 
\code{\link{dec2base}} converts decimal numbers into numerals in another base.

Other numeric functions: 
\code{\link{base2dec}()},
\code{\link{dec2base}()},
\code{\link{is_equal}()},
\code{\link{is_wholenumber}()},
\code{\link{num_as_char}()},
\code{\link{num_as_ordinal}()},
\code{\link{num_equal}()}

Other utility functions: 
\code{\link{base2dec}()},
\code{\link{dec2base}()},
\code{\link{is_equal}()},
\code{\link{is_vect}()},
\code{\link{is_wholenumber}()},
\code{\link{num_as_char}()},
\code{\link{num_as_ordinal}()},
\code{\link{num_equal}()}
}
\concept{numeric functions}
\concept{utility functions}
\keyword{datasets}
