% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{caseflip}
\alias{caseflip}
\title{caseflip flips the case of characters 
in a string of text \code{x}.}
\usage{
caseflip(x)
}
\arguments{
\item{x}{A string of text (required).}
}
\description{
caseflip flips the case of characters 
in a string of text \code{x}.
}
\examples{
x <- c("Hello world!", "This is a 1st sentence.", "This is the 2nd sentence.", "The end.")
caseflip(x)
 
}
\seealso{
\code{\link{capitalize}} for converting the case of initial letters.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{count_char}()},
\code{\link{l33t_rul35}},
\code{\link{metachar}},
\code{\link{read_ascii}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
