% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{count_word}
\alias{count_word}
\title{count_word counts the frequency of words  
in a string of text \code{x}.}
\usage{
count_word(x, case_sense = TRUE, sort_freq = TRUE)
}
\arguments{
\item{x}{A string of text (required).}

\item{case_sense}{Boolean: Distinguish lower- vs. uppercase characters? 
Default: \code{case_sense = TRUE}.}

\item{sort_freq}{Boolean: Sort output by word frequency? 
Default: \code{sort_freq = TRUE}.}
}
\description{
count_word counts the frequency of words  
in a string of text \code{x}.
}
\examples{
# Default: 
s3 <- c("A first sentence.", "The second sentence.", 
        "A third --- and also the final --- sentence.")
count_word(s3)  # case-sensitive, sorts by frequency 

# Options: 
count_word(s3, case_sense = FALSE)  # case insensitive
count_word(s3, sort_freq = FALSE)   # sorts alphabetically
 
}
\seealso{
\code{\link{count_char}} for counting the frequency of characters;   
\code{\link{plot_text}} for a corresponding plot function.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{count_char}()},
\code{\link{l33t_rul35}},
\code{\link{metachar}},
\code{\link{read_ascii}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
