% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_1}
\alias{data_1}
\title{Data import data_1.}
\format{
A tibble with 100 cases (rows) and 4 variables (columns).
}
\source{
See DELIMITED data at \url{http://rpository.com/ds4psy/data/data_1.dat}.
}
\usage{
data_1
}
\description{
\code{data_1} is a fictitious dataset to practice data import 
(from a DELIMITED file).
}
\seealso{
Other datasets: 
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
