% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t1_tab}
\alias{data_t1_tab}
\title{Data import data_t1_tab.}
\format{
A table with 20 cases (rows) and 4 variables (columns).
}
\source{
See TAB-delimited data at \url{http://rpository.com/ds4psy/data/data_t1_tab.csv}.
}
\usage{
data_t1_tab
}
\description{
\code{data_t1_tab} is a fictitious dataset to practice data import 
(from a TAB file).
}
\seealso{
Other datasets: 
\code{\link{Bushisms}},
\code{\link{Trumpisms}},
\code{\link{countries}},
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{flowery}},
\code{\link{fruits}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{t_1}},
\code{\link{t_2}},
\code{\link{t_3}},
\code{\link{t_4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
