% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t3}
\alias{data_t3}
\title{Data table data_t3.}
\format{A tibble with 20 cases (rows) and 4 variables (columns).}
\source{
See CSV data at \url{http://rpository.com/ds4psy/data/data_t3.csv}.
}
\usage{
data_t3
}
\description{
\code{data_t3} is a fictitious dataset to practice importing and joining data 
(from a CSV file).
}
\seealso{
Other datasets: \code{\link{data_1}}, \code{\link{data_2}},
  \code{\link{data_t1_de}}, \code{\link{data_t1_tab}},
  \code{\link{data_t1}}, \code{\link{data_t2}},
  \code{\link{data_t4}}, \code{\link{exp_wide}},
  \code{\link{falsePosPsy_all}}, \code{\link{fame}},
  \code{\link{outliers}}, \code{\link{pi_100k}},
  \code{\link{posPsy_AHI_CESD}}, \code{\link{posPsy_long}},
  \code{\link{posPsy_p_info}}, \code{\link{posPsy_wide}},
  \code{\link{t3}}, \code{\link{t4}}, \code{\link{table6}},
  \code{\link{table7}}, \code{\link{table8}},
  \code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
