% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_fun.R
\name{dec2base}
\alias{dec2base}
\title{Convert an integer in decimal notation into a string of numeric digits in some base.}
\usage{
dec2base(x, base = 2, as_char = TRUE)
}
\arguments{
\item{x}{A (required) integer in decimal (base 10) notation 
or corresponding string of digits (i.e., 0-9).}

\item{base}{The base or radix of the digits in the output. 
Default: \code{base = 2} (binary).}

\item{as_char}{Return the output as a character string? 
Default: \code{as_char = TRUE} (as symbol sequence is NOT a 
decimal number unless \code{base = 10}).}
}
\value{
A string of digits (in base notation).
}
\description{
\code{dec2base} converts an integer from decimal 
(i.e., base or radix 10) notation 
into a sequence of numeric symbols (digits) of some other base.
}
\details{
To prevent erroneous interpretations of numeric outputs, 
\code{dec2base} returns a sequence of digits (as a character string).
When using \code{as_char = FALSE}, its output string is 
processed by \code{as.integer}, but this may cause 
problems with the interpretation of the numeric value 
(as outputs for a base/radix other than 10 do NOT denote decimal numbers) 
and scientific notation. 

\code{dec2base} is the complement of \code{\link{base2dec}}.
}
\examples{
# (a) single numeric input:
dec2base(3)  # base = 2
dec2base(4)
dec2base(8)

dec2base(8, base = 3)
dec2base(8, base = 7)

dec2base(100, base = 2)
dec2base(100, base = 5)
dec2base(100, base = 10)

# (b) single string input:
dec2base("7", base = 2)
dec2base("8", base = 3)

# Note: 
base2dec(dec2base(012340, base = 5), base = 5)
dec2base(base2dec(043210, base = 5), base = 5)

}
\seealso{
\code{\link{base2dec}} converts numerals in some base into decimal numbers; 
\code{\link{as.roman}} converts integers into Roman numerals.

Other numeric functions: 
\code{\link{base2dec}()},
\code{\link{base_digits}},
\code{\link{is_equal}()},
\code{\link{is_wholenumber}()},
\code{\link{num_as_char}()},
\code{\link{num_as_ordinal}()},
\code{\link{num_equal}()}

Other utility functions: 
\code{\link{base2dec}()},
\code{\link{base_digits}},
\code{\link{is_equal}()},
\code{\link{is_vect}()},
\code{\link{is_wholenumber}()},
\code{\link{num_as_char}()},
\code{\link{num_as_ordinal}()},
\code{\link{num_equal}()}
}
\concept{numeric functions}
\concept{utility functions}
