% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun.R
\name{dice}
\alias{dice}
\title{Throw a fair dice (with a given number of sides) n times.}
\usage{
dice(n = 1, sides = 6)
}
\arguments{
\item{n}{Number of dice throws.
Default: \code{n = 1}.}

\item{sides}{Number of sides.
Default: \code{sides = 6}.}
}
\description{
\code{dice} generates a sequence of events that 
represent the results of throwing a fair dice 
(with a given number of \code{sides}) \code{n} times.
}
\examples{
# Basics:
dice()
table(dice(10^4))

# 5-sided dice:
dice(sides = 5)
table(dice(100, sides = 5))

# Note:
dice(10, 1)
table(dice(1000, 2))

# Limits:
dice(NA)
dice(0)
dice(1/2)
dice(1:2)
dice(10, sides = NA)
dice(10, sides = 1/2)
dice(10, sides = 1:2)

}
\seealso{
Other random functions: \code{\link{coin}},
  \code{\link{dice_2}}
}
\concept{random functions}
