% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{diff_dates}
\alias{diff_dates}
\title{Get the difference between two dates (in human units).}
\usage{
diff_dates(from_date, to_date = Sys.Date(), unit = "y", as_character = TRUE)
}
\arguments{
\item{from_date}{From date (required, scalar or vector, as "Date"). 
Date of birth (DOB), assumed to be of class "Date", 
and coerced into "Date" when of class "POSIXt".}

\item{to_date}{To date (optional, scalar or vector, as "Date"). 
Default: \code{to_date = Sys.Date()}. 
Maximum date/date of death (DOD), assumed to be of class "Date", 
and coerced into "Date" when of class "POSIXt".}

\item{unit}{Largest measurement unit to represent output (as "character").
Units represent human time periods, rather than 
chronological time differences. 
Default: \code{unit = "y"} for completed years, months, and days. 
Available options are: 

\enumerate{

  \item \code{unit = "y"}: completed years, months, and days (default)
  
  \item \code{unit = "m"}: completed months, and days
  
  \item \code{unit = "d"}: completed days
  
  }}

\item{as_character}{Boolean: Return output as character? 
Default: \code{as_character = TRUE}.  
If \code{as_character = FALSE}, output is returned 
as numeric columns of a data frame.}
}
\description{
\code{diff_dates} computes the difference between two dates 
(i.e., from some \code{from_date} to some \code{to_date}) 
in human measurement units (periods).
}
\details{
\code{diff_dates} answers questions like  
"How much time has elapsed between two dates?" 
or "How old are you?" 
in human units of (full) years, months, or days. 

If not specified explicitly, \code{to_date} is set to 
today's date (i.e., \code{Sys.Date()}).

If the lengths of \code{from_date} and \code{to_date} differ, 
the arguments of \code{to_date} are recycled or 
truncated to the length of \code{from_date}.
}
\examples{
y_100 <- Sys.Date() - (100 * 365.25) + -1:1
diff_dates(y_100)

# with "to_date" argument: 
y_050 <- Sys.Date() - (50 * 365.25) + -1:1 
diff_dates(y_100, y_050)

# robustness:
days_this_year <- 365 + is_leap_year(Sys.Date())
diff_dates(Sys.time() - (10 * (60 * 60 * 24) * days_this_year)) # for POSIXt times
diff_dates("90-07-11", to_date = "10-07-10")                    # for strings
diff_dates(19900711, to_date = 20100710)                        # for numbers

# recycling "to_date" to length of "from_date":
y_050_2 <- Sys.Date() - (50 * 365.25)
diff_dates(y_100, y_050_2)

# Using 'fame' data:
dob <- as.Date(fame$DOB, format = "\%B \%d, \%Y")
dod <- as.Date(fame$DOD, format = "\%B \%d, \%Y")
diff_dates(dob, dod)  # Note: Deceased people do not age further.

}
\seealso{
Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{is_leap_year}()},
\code{\link{what_date}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_wday}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
