% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_util_fun.R
\name{diff_tz}
\alias{diff_tz}
\title{Get the time zone difference between two times.}
\usage{
diff_tz(t1, t2, in_min = FALSE)
}
\arguments{
\item{t1}{1st time point (required, as "POSIXt").}

\item{t2}{2nd time point (required, as "POSIXt").}

\item{in_min}{Return time-zone based time 
difference in minutes (Boolean)? 
Default: \code{in_min = FALSE}.}
}
\value{
A character (in "HH:MM" format) or 
numeric vector (number of minutes).
}
\description{
\code{diff_tz} computes the time difference 
between two times \code{t1} and \code{t2} 
that is due to both times being in different time zones.
}
\details{
\code{diff_tz} allows adjusting time-based computations 
for shifts that are entirely due to time zone differences,  
rather than differences in actual times. 

Internally, \code{diff_tz} uses and compares the POSIX 
conversion specifications "%Z" and "%z" (in numeric form). 

If the lengths of \code{t1} and \code{t2} differ, 
the arguments of \code{t2} are recycled or 
truncated to the length of \code{t1}.
}
\examples{
# Time zones differences:
tm <- "2020-01-01 01:00:00"  # nominal time
t1 <- as.POSIXct(tm, tz = "NZ")
t2 <- as.POSIXct(tm, tz = "Europe/Berlin")
t3 <- as.POSIXct(tm, tz = "US/Hawaii")

# as character (in "HH:MM"):
diff_tz(t1, t2)
diff_tz(t2, t3)
diff_tz(t1, t3)

# as numeric (in minutes):
diff_tz(t1, t2, in_min = TRUE)
diff_tz(t2, t3, in_min = TRUE)
diff_tz(t1, t3, in_min = TRUE)

# Compare local times (POSIXlt): 
t1 <- as.POSIXlt(Sys.time(), tz = "NZ")
t2 <- as.POSIXlt(Sys.time(), tz = "Europe/Berlin")
diff_tz(t1, t2)

# DSL shift: Spring ahead (on 2020-03-29: 02:00:00 > 03:00:00):
s1 <- "2020-03-29 01:00:00 CET"   # before DSL switch
s2 <- "2020-03-29 03:00:00 CEST"  # after DSL switch
t1 <- as.POSIXct(s1, tz = "Europe/Berlin")  # CET
t2 <- as.POSIXct(s2, tz = "Europe/Berlin")  # CEST

diff_tz(t1, t2)
diff_tz(t1, t2, in_min = TRUE)

}
\seealso{
\code{\link{days_in_month}} for the number of days in given months; 
\code{\link{is_leap_year}} to check for leap years.

Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{days_in_month}()},
\code{\link{diff_dates}()},
\code{\link{diff_times}()},
\code{\link{is_leap_year}()},
\code{\link{what_date}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_wday}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
