% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{ds4psy.guide}
\alias{ds4psy.guide}
\title{Open the \strong{ds4psy} package's user guide.}
\usage{
ds4psy.guide()
}
\description{
The \strong{ds4psy} package currently only contains a default vignette 
that provides general information and links.
}
