% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruits}
\alias{fruits}
\title{Data: Names of fruits.}
\format{
A vector of type \code{character}  
with \code{length(fruits) = 122}.
}
\source{
Data based on \url{https://simple.wikipedia.org/wiki/List_of_fruits}.
}
\usage{
fruits
}
\description{
\code{fruits} is a dataset containing the names of 
122 fruits (as a vector of text strings).
}
\details{
Botanically, "fruits" are the seed-bearing structures 
of flowering plants (angiosperms) formed from the ovary 
after flowering. 

In common usage, "fruits" refer to the fleshy 
seed-associated structures of a plant 
that taste sweet or sour, 
and are edible in their raw state.
}
\seealso{
Other datasets: 
\code{\link{Bushisms}},
\code{\link{Trumpisms}},
\code{\link{countries}},
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{dt_10}},
\code{\link{exp_num_dt}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{flowery}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{t_1}},
\code{\link{t_2}},
\code{\link{t_3}},
\code{\link{t_4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{table9}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
