% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun.R
\name{get_set}
\alias{get_set}
\title{Get a set of x-y coordinates.}
\source{
See \code{?datasets:anscombe} for details and references.
}
\usage{
get_set(n = 1)
}
\arguments{
\item{n}{Number of set (as an integer).
Default: \code{n = 1}.}
}
\description{
\code{get_set} obtains a set of x/y coordinates and returns it 
(as a data frame).
}
\details{
The set stems from Anscombe's Quartet 
(hence \code{1 <= 0 <= 4}) and is returned as a 
data frame (with 11 rows and 2 columns).
}
\examples{
get_set(1)
plot(get_set(2), col = "red")

}
\seealso{
Other data functions: 
\code{\link{make_grid}()}
}
\concept{data functions}
