% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\docType{data}
\name{l33t_rul35}
\alias{l33t_rul35}
\title{l33t_rul35 provides rules for translating into leet/l33t slang.}
\format{An object of class \code{character} of length 13.}
\usage{
l33t_rul35
}
\description{
\code{l33t_rul35} specifies rules for translating characters 
into other characters (typically symbols) to mimic 
leet/l33t slang (as a named character vector).
}
\details{
Old (i.e., to be replaced) characters are 
\code{paste(names(l33t_rul35), collapse = "")}.

New (i.e., replaced) characters are 
\code{paste(l33t_rul35, collapse = "")}.

See \url{https://en.wikipedia.org/wiki/Leet} for details.
}
\seealso{
Other text functions: 
\code{\link{count_char}()},
\code{\link{read_ascii}()},
\code{\link{transl33t}()}
}
\concept{text functions}
\keyword{datasets}
