% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun.R
\name{make_grid}
\alias{make_grid}
\title{Generate a grid of x-y coordinates.}
\usage{
make_grid(x_min = 0, x_max = 2, y_min = 0, y_max = 1)
}
\arguments{
\item{x_min}{Minimum x coordinate.
Default: \code{x_min = 0}.}

\item{x_max}{Maximum x coordinate.
Default: \code{x_max = 2}.}

\item{y_min}{Minimum y coordinate.
Default: \code{y_min = 0}.}

\item{y_max}{Maximum y coordinate.
Default: \code{y_max = 1}.}
}
\description{
\code{make_grid} generates a grid of x/y coordinates and returns it 
(as a data frame).
}
\examples{
make_grid()
make_grid(x_min = -3, x_max = 3, y_min = -2, y_max = 2)

}
\concept{data functions}
