% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\docType{data}
\name{metachar}
\alias{metachar}
\title{metachar provides R metacharacters 
(as a character vector).}
\format{
An object of class \code{character} of length 12.
}
\usage{
metachar
}
\description{
\code{metachar} provides the metacharacters of extended regular expressions 
(as a character vector).
}
\details{
See \code{?regex} for details.
}
\examples{
metachar
length(metachar)  # 12
nchar(paste0(metachar, collapse = ""))  # 12

}
\seealso{
Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{count_char}()},
\code{\link{count_word}()},
\code{\link{l33t_rul35}},
\code{\link{read_ascii}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
\keyword{datasets}
