% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outliers}
\alias{outliers}
\title{Outlier data.}
\format{
A table with 100 cases (rows) and 3 variables (columns).
}
\source{
See CSV data at \url{http://rpository.com/ds4psy/data/out.csv}.
}
\usage{
outliers
}
\description{
\code{outliers} is a fictitious dataset containing the id, sex, and height 
of 1000 non-existing, but otherwise normal people.
}
\details{
\strong{Codebook}

\describe{
  \item{id}{Participant ID (as character code)}
  \item{sex}{Gender (female vs. male)}
  \item{height}{Height (in cm)}
}
}
\seealso{
Other datasets: 
\code{\link{Bushisms}},
\code{\link{Trumpisms}},
\code{\link{countries}},
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{flowery}},
\code{\link{fruits}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{t_1}},
\code{\link{t_2}},
\code{\link{t_3}},
\code{\link{t_4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
