% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_fun.R
\docType{data}
\name{pal_ds4psy}
\alias{pal_ds4psy}
\title{ds4psy default color palette.}
\format{
An object of class \code{data.frame} with 1 rows and 11 columns.
}
\usage{
pal_ds4psy
}
\description{
\code{pal_ds4psy} provides a dedicated color palette.
}
\details{
By default, \code{pal_ds4psy} is based on   
\code{pal_unikn} of the \bold{unikn} package.
}
\seealso{
Other color objects and functions: 
\code{\link{pal_n_sq}()}
}
\concept{color objects and functions}
\keyword{datasets}
