% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pal_n_sq}
\alias{pal_n_sq}
\title{Get n^2 dedicated colors of a color palette.}
\usage{
pal_n_sq(n = "all", pal = pal_ds4psy)
}
\arguments{
\item{n}{A number specifying the desired number colors of pal (as a number) 
or the character string \code{"all"} (to get all colors of \code{pal}). 
Default: \code{n = "all"}.}

\item{pal}{A color palette (as a data frame). 
Default: \code{pal = \link{pal_ds4psy}}.}
}
\description{
\code{pal_n_sq} returns \code{n^2} dedicated colors of a color palette \code{pal} 
(up to a maximum of \code{n = "all"} colors).
}
\details{
Note that \code{pal_n_sq} was originally created for \code{pal = \link{pal_unikn}} 
for small values of \code{n} (\code{n = 1, 2, 3}) and 
returned the 11 colors of \code{\link{pal_unikn_plus}} for any \code{n > 3}.

Use the more specialized function \code{unikn::usecol} for choosing 
\code{n} dedicated colors of a known color palette.
}
\examples{
pal_n_sq(1)  #  1 color: seeblau3
pal_n_sq(2)  #  4 colors
pal_n_sq(3)  #  9 colors (5: white)
pal_n_sq(4)  # 11 colors (6: white)

}
\seealso{
\code{\link{unikn::seecol}} to plot color palettes; 
\code{\link{unikn::usecol}} to use color palettes; 
\code{\link{unikn::pal_unikn}} for the default unikn color palette.

Other plot functions: \code{\link{pal_ds4psy}},
  \code{\link{theme_ds4psy}}
}
\concept{plot functions}
