% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pi_100k}
\alias{pi_100k}
\title{Data: 100k digits of pi.}
\format{A character of \code{nchar(pi_100k) = 100001}.}
\source{
See TXT data at \url{http://rpository.com/ds4psy/data/pi_100k.txt}. 

Original data at \url{http://www.geom.uiuc.edu/~huberty/math5337/groupe/digits.html}.
}
\usage{
pi_100k
}
\description{
A dataset containing the first 100k digits of pi.
}
\seealso{
Other datasets: \code{\link{data_1}}, \code{\link{data_2}},
  \code{\link{data_t1_de}}, \code{\link{data_t1_tab}},
  \code{\link{data_t1}}, \code{\link{data_t2}},
  \code{\link{data_t3}}, \code{\link{data_t4}},
  \code{\link{exp_wide}}, \code{\link{falsePosPsy_all}},
  \code{\link{outliers}}, \code{\link{posPsy_AHI_CESD}},
  \code{\link{posPsy_long}}, \code{\link{posPsy_p_info}},
  \code{\link{posPsy_wide}}, \code{\link{t3}},
  \code{\link{t4}}, \code{\link{table6}},
  \code{\link{table7}}, \code{\link{table8}},
  \code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
