% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\name{plot_circ_points}
\alias{plot_circ_points}
\title{Plot objects (as points) arranged on a circle}
\usage{
plot_circ_points(
  n_points = 4,
  x_org = 0,
  y_org = 0,
  radius = 1,
  show_label = FALSE,
  ...
)
}
\arguments{
\item{n_points}{The number of points to plot}

\item{x_org}{The x-value of circle origin}

\item{y_org}{The y-value of circle origin}

\item{radius}{The circle radius}

\item{show_label}{Show a point label? Default: \code{show_label = FALSE}.}

\item{...}{Additional aesthetics (passed to \code{\link{points}} of \strong{graphics}).}
}
\description{
\code{plot_circ_points} arranges a number of \code{n_points} 
on a circle (defined by its origin coordinates and radius).
}
\details{
The \code{...} is passed to \code{\link{points}} of 
the \strong{graphics} package.
}
\examples{
plot_circ_points()  # default
plot_circ_points(12, show_label = TRUE, cex = 5, pch = 21, col = "grey", bg = "gold")
 
}
