% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\name{plot_fn}
\alias{plot_fn}
\title{A function to plot a plot.}
\usage{
plot_fn(x = NA, y = 1, A = TRUE, B = FALSE, C = TRUE,
  D = FALSE, E = FALSE, F = FALSE, f = c(rev(pal_seeblau), "white",
  Grau, "black", Petrol, Pinky), g = "white")
}
\arguments{
\item{x}{A (natural) number. 
Default: \code{x = NA}.}

\item{y}{A (decimal) number. 
Default: \code{y = 0}.}

\item{A}{A Boolean value. 
Default: \code{A = TRUE}.}

\item{B}{A Boolean value. 
Default: \code{B = FALSE}.}

\item{C}{A Boolean value. 
Default: \code{C = TRUE}.}

\item{D}{A Boolean value. 
Default: \code{D = FALSE}.}

\item{E}{A Boolean value. 
Default: \code{E = FALSE}.}

\item{F}{A Boolean value. 
Default: \code{F = FALSE}.}

\item{f}{A color palette (e.g., as a vector). 
Default: \code{f = c(rev(pal_seeblau), "white", Grau, "black", Petrol, Pinky)}. 
Note: Using colors of the \code{unikn} package by default.}

\item{g}{A color (e.g., as a character). 
Default: \code{g = "white"}.}
}
\description{
\code{plot_fn} is a function that uses parameters for plotting a plot.
}
\details{
\code{plot_fn} is deliberately kept cryptic and obscure to illustrate 
how function parameters can be explored --- and why transparent variable 
names are essential for understanding and using a function.
}
\examples{
# Basics: 
plot_fn()

# Exploring options: 
plot_fn(x = 2, A = TRUE)
plot_fn(x = 3, A = FALSE)
plot_fn(x = 4, A = TRUE,  B = TRUE, D = TRUE)
plot_fn(x = 5, A = FALSE, B = TRUE, E = TRUE)
plot_fn(x = 6, A = FALSE, B = TRUE, F = TRUE, f = c("steelblue", "white", "forestgreen"))

}
\seealso{
\code{\link{plot_fun}} for a related function; 
\code{\link{pal_ds4psy}} for color palette.

Other plot functions: \code{\link{plot_fun}},
  \code{\link{plot_n}}, \code{\link{plot_tiles}},
  \code{\link{theme_ds4psy}}
}
\concept{plot functions}
