% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\name{plot_txt}
\alias{plot_txt}
\title{Plot text characters (from file).}
\usage{
plot_txt(file = "txt/hello.txt", lbl_tiles = TRUE, cex = 3,
  fontface = 1, col_txt = "black", col_bg = "white",
  pal = pal_ds4psy[1:5], pal_extend = TRUE, case_sense = FALSE,
  borders = TRUE, border_col = "white", border_size = 0.5)
}
\arguments{
\item{file}{The text file to read (or its path).
If the text file is stored in a sub-directory, 
enter its path and name here (without any leading or 
trailing "." or "/"). 
Default: \code{file = "txt/hello.txt"}.}

\item{lbl_tiles}{Add numeric labels to tiles? 
Default: \code{lbl_tiles = TRUE} (i.e., show labels).}

\item{cex}{Character size (numeric). 
Default: \code{cex = 3}.}

\item{fontface}{Font face (numeric). 
Default: \code{fontface = 1}, (from 1 to 4).}

\item{col_txt}{Color of text characters.
Default: \code{col_txt = "black"} (if \code{lbl_tiles = TRUE}).}

\item{col_bg}{Color of most frequent character in text 
(typically " ", i.e., background). 
Default: \code{col_bg = "white"}.}

\item{pal}{Color palette to use for tiles 
of text (used in order of character frequency). 
Default: \code{pal = pal_ds4psy[1:5]}.}

\item{pal_extend}{Boolean: Should pal be extended 
to match the number of different characters in text? 
Default: \code{pal_extend = TRUE}. 
If \code{pal_extend = FALSE}, only the tiles of 
the \code{length(pal)} most frequent characters 
will be filled by the colors of \code{pal}.}

\item{case_sense}{Boolean: Should lower- and 
uppercase characters be distinguished? 
Default: \code{case_sense = FALSE}.}

\item{borders}{Boolean: Add borders to tiles? 
Default: \code{borders = TRUE} (i.e., use borders).}

\item{border_col}{Color of borders (if \code{borders = TRUE}). 
Default: \code{border_col = "white"}.}

\item{border_size}{Size of borders (if \code{borders = TRUE}). 
Default: \code{border_size = 0.5}.}
}
\description{
\code{plot_txt} parses text (from a file) 
into a tibble and then plots all 
its characters as a tile plot (using \strong{ggplot2}).
}
\examples{

\donttest{
plot_txt("txt/hello.txt")  # requires txt file

# Colors, pal_extend, and case_sense:
cols <- c("firebrick", "olivedrab", "steelblue", "orange", "gold")
plot_txt(pal = cols, pal_extend = TRUE)
plot_txt(pal = cols, pal_extend = FALSE)
plot_txt(pal = cols, pal_extend = FALSE, case_sense = TRUE)

# Customize:
plot_txt(col_txt = "white", borders = FALSE)
plot_txt(col_txt = "white", pal = c("green4", "black"),
         border_col = "black", border_size = .2)

# Color ranges:
plot_txt(pal = c("red2", "orange", "gold"))
plot_txt(pal = c("olivedrab4", "gold"))

# Text and grid options:
plot_txt(col_txt = "firebrick", cex = 4, fontface = 3,
         pal = "grey90", pal_extend = TRUE,
         border_col = NA)

# Other text file:
plot_txt(file = "txt/ascii.txt", cex = 5, 
         col_bg = "lightgrey", border_col = "white")
}

}
\seealso{
\code{\link{read_ascii}} for reading text files into a tibble; 
\code{\link{pal_ds4psy}} for default color palette.

Other plot functions: \code{\link{plot_fn}},
  \code{\link{plot_fun}}, \code{\link{plot_n}},
  \code{\link{plot_tiles}}, \code{\link{theme_ds4psy}}
}
\concept{plot functions}
