% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posPsy_AHI_CESD}
\alias{posPsy_AHI_CESD}
\title{Positive Psychology: AHI_CESD data.}
\format{A tibble with 992 cases (rows) and 50 variables (columns).
 
\strong{Codebook} 

1. id: Particpant ID. 

2. occasion: Measurement occasion:  
  0 = Pretest (i.e., at enrolment),
  1 = Posttest (i.e., 7 days after pretest),
  2 = 1-week follow-up, (i.e., 14 days after pretest, 7 days after posttest),
  3 = 1-month follow-up, (i.e., 38 days after pretest, 31 days after posttest),
  4 = 3-month follow-up, (i.e., 98 days after pretest, 91 days after posttest),
  5 = 6-month follow-up, (i.e., 189 days after pretest, 182 days after posttest).

3. elapsed.days: Time since enrolment measured in fractional days. 

4. intervention: Intervention group (1 to 4). 

- Block ahi01–ahi24: Responses on 24 AHI items. 

- Block cesd01–cesd20: Responses on 20 CES-D items. 

49. ahiTotal: Total AHI score. 

50. cesdTotal: Total CES-D score.}
\source{
\strong{Articles}

Woodworth, R. J., O’Brien‐Malone, A., Diamond, M. R., & Schüz, B. (2017). 
Web‐based positive psychology interventions: A reexamination of effectiveness. 
\emph{Journal of Clinical Psychology}, \emph{73}(3), 218–232. 
doi: \code{10.1002/jclp.22328} 

Woodworth, R. J., O’Brien-Malone, A., Diamond, M. R. and Schüz, B. (2018). 
Data from, ‘Web-based positive psychology interventions: A reexamination of effectiveness’. 
\emph{Journal of Open Psychology Data}, \emph{6}(1). 
doi: \code{10.5334/jopd.35} 

\strong{Data}

See details at \url{https://doi.org/10.6084/m9.figshare.1577563.v1}.
}
\usage{
posPsy_AHI_CESD
}
\description{
This dataset contains answers to the 24 items of the 
Authentic Happiness Inventory (AHI) and answers to the 
20 items of the Center for Epidemiological Studies Depression (CES-D) scale 
(see Radloff, 1977) for multiple (1 to 6) measurement occasions.
}
\seealso{
\code{\link{posPsy_long}} for a corrected version of this file (in long format).

Other datasets: \code{\link{falsePosPsy_all}},
  \code{\link{posPsy_long}}, \code{\link{posPsy_p_info}},
  \code{\link{posPsy_wide}}
}
\concept{datasets}
\keyword{datasets}
