% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posPsy_p_info}
\alias{posPsy_p_info}
\title{Positive Psychology: Participant data.}
\format{A tibble with 295 cases (rows) and 6 variables (columns):
\describe{
  \item{id}{Participant ID.}
  \item{intervention}{Type of intervention: 1-3: PPI, 4: control condition.}
  \item{sex}{Sex: 1 = female, 2 = male.}
  \item{age}{Age in years.}
  \item{educ}{Education level: 1: less than 12 years, 5: Postgraduate degree.}
  \item{income}{Income: 1: below average, 3: above average.}
}}
\source{
Woodworth, R. J., O’Brien‐Malone, A., Diamond, M. R., & Schüz, B. (2017). 
Web‐based positive psychology interventions: A reexamination of effectiveness. 
Journal of Clinical Psychology, 73(3), 218–232. doi: 10.1002/jclp.22328

Woodworth, R. J., O’Brien-Malone, A., Diamond, M. R. and Schüz, B. (2018). 
Data from, ‘Web-based positive psychology interventions: A reexamination of effectiveness’. 
Journal of Open Psychology Data, 6: 1. doi: 10.5334/jopd.35

See details at \url{https://doi.org/10.6084/m9.figshare.1577563.v1}.
}
\usage{
posPsy_p_info
}
\description{
A dataset containing details of 295 participants.
}
\keyword{datasets}
