% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posPsy_wide}
\alias{posPsy_wide}
\title{Positive Psychology: All corrected data (in wide format).}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 295 rows and 294 columns.}
\source{
Woodworth, R. J., O’Brien‐Malone, A., Diamond, M. R., & Schüz, B. (2017). 
Web‐based positive psychology interventions: A reexamination of effectiveness. 
Journal of Clinical Psychology, 73(3), 218–232. doi: 10.1002/jclp.22328

Woodworth, R. J., O’Brien-Malone, A., Diamond, M. R. and Schüz, B. (2018). 
Data from, ‘Web-based positive psychology interventions: A reexamination of effectiveness’. 
Journal of Open Psychology Data, 6: 1. doi: 10.5334/jopd.35

See details at \url{https://doi.org/10.6084/m9.figshare.1577563.v1}.
}
\usage{
posPsy_wide
}
\description{
Positive Psychology: All corrected data (in wide format).
}
\keyword{datasets}
