% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{read_ascii}
\alias{read_ascii}
\title{read_ascii parses text (from a file) into a tibble.}
\usage{
read_ascii(file = "data-raw/ascii.txt")
}
\arguments{
\item{file}{The text file to read or its path.
If the text file is stored in a sub-directory, 
enter this here (without leading or trailing "/"). 
Default: \code{file = "data-raw/ascii.txt"}.}
}
\description{
\code{read_ascii} parses text (from a file) into 
a tibble that contains a row for each character. 
This tibble contains 3 variables: 
The character's \code{x}- and \code{y}-coordinates (from top to bottom)  
and a variable \code{char} for the character at this coordinate.
}
\details{
The \bold{here} package is used to determine 
the (absolute) file path.
}
\examples{
# depend on file.txt

}
\seealso{
corresponding plot function

Other text functions: \code{\link{l33t_rul35}},
  \code{\link{transl33t}}
}
\concept{text functions}
