% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t3}
\alias{t3}
\title{Data table t3.}
\format{A tibble with 10 cases (rows) and 4 variables (columns).}
\source{
See CSV data at \url{http://rpository.com/ds4psy/data/t3.csv}.
}
\usage{
t3
}
\description{
A ficticious dataset to practice importing and joining data (from CSV file).
}
\seealso{
Other datasets: \code{\link{data_1}}, \code{\link{data_2}},
  \code{\link{data_t1_de}}, \code{\link{data_t1_tab}},
  \code{\link{data_t1}}, \code{\link{data_t2}},
  \code{\link{data_t3}}, \code{\link{data_t4}},
  \code{\link{exp_wide}}, \code{\link{falsePosPsy_all}},
  \code{\link{outliers}}, \code{\link{pi_100k}},
  \code{\link{posPsy_AHI_CESD}}, \code{\link{posPsy_long}},
  \code{\link{posPsy_p_info}}, \code{\link{posPsy_wide}},
  \code{\link{t4}}, \code{\link{table6}},
  \code{\link{table7}}, \code{\link{table8}},
  \code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
