% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table8}
\alias{table8}
\title{Data table8.}
\format{A tibble with 3 cases (rows) and 5 variables (columns).}
\source{
See CSV data at \url{http://rpository.com/ds4psy/data/table8.csv}.
}
\usage{
table8
}
\description{
\code{table8} is a fictitious dataset to practice tidying data.
}
\details{
This dataset is a variant of the \code{table1} to \code{table5} datasets 
of the \bold{tidyr} package.
}
\seealso{
Other datasets: 
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
