% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{text_to_words}
\alias{text_to_words}
\title{text_to_words turns a string of text \code{x} 
(consisting of one or more character strings) 
into a vector of its words.}
\usage{
text_to_words(x)
}
\arguments{
\item{x}{A string of text (required).}
}
\description{
text_to_words first removes all punctuation and empty spaces 
and returns a vector of all remaining character symbols 
(as the words).
}
\examples{
# Default: 
x <- c("Hello!", "This is a 1st sentence.", "This is the 2nd sentence.", "The end.")
text_to_words(x)

}
\seealso{
\code{\link{count_word}} for counting the frequency of words.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{count_char}()},
\code{\link{count_word}()},
\code{\link{l33t_rul35}},
\code{\link{metachar}},
\code{\link{read_ascii}()},
\code{\link{text_to_sentences}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
