% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{what_age}
\alias{what_age}
\title{What is the age (in full years) between from and to date?}
\usage{
what_age(from_date, to_date = Sys.Date())
}
\arguments{
\item{from_date}{Date (required, as scalar or vector). 
Date of birth (DOB), assumed to be of class "Date", 
and co-erced into "Date" when of class "POSIXt".}

\item{to_date}{Date (optional, as scalar or vector). 
Default: \code{to_date = Sys.Date()}. 
Maximum date/date of death (DOD), assumed to be of class "Date", 
and co-erced into "Date" when of class "POSIXt".}
}
\description{
\code{what_age} provides the number of completed years 
from some \code{from_date} to some \code{to_date}.
}
\details{
If the lengths of \code{from_date} and \code{to_date} differ, 
the arguments of \code{to_date} are recycled or 
truncated to the length of \code{from_date}.
}
\examples{
y_100 <- Sys.Date() - (100 * 365.25) + -1:1
what_age(y_100)

# with "to_date" argument: 
y_050 <- Sys.Date() - (50 * 365.25) + -1:1 
what_age(y_100, y_050)

# recycling "to_date" to length of "from_date":
y_050_2 <- Sys.Date() - (50 * 365.25)
what_age(y_100, y_050_2)

# Using 'fame' data:
dob <- as.Date(fame$DOB, format = "\%B \%d, \%Y")
dod <- as.Date(fame$DOD, format = "\%B \%d, \%Y")
what_age(dob, dod)

}
\seealso{
Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{is_leap_year}()},
\code{\link{what_date}()},
\code{\link{what_day}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
