% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{what_day_alt}
\alias{what_day_alt}
\title{What day is it? (alternative version)}
\usage{
what_day_alt(
  when = Sys.time(),
  unit = "week",
  abbr = FALSE,
  as_integer = FALSE
)
}
\arguments{
\item{when}{Date (as a scalar or vector).
Default: \code{when = NA}. 
Using \code{as.Date(when)} to convert strings into dates, 
and \code{Sys.Date()}, if \code{when = NA}.}

\item{unit}{Character: Unit of day?
Possible values are \code{"week", "month", "year"}. 
Default: \code{unit = "week"} (for day within week).}

\item{abbr}{Boolean: Return abbreviated?
Default: \code{abbr = FALSE}.}

\item{as_integer}{Boolean: Return as integer? 
Default: \code{as_integer = FALSE}.}
}
\description{
\code{what_day_alt} provides a satisficing version of 
to determine the day corresponding to a given date.
}
\details{
\code{what_day_alt} returns the day  
of \code{when} or \code{Sys.Date()} 
(as a name or number).
}
\examples{
what_day_alt()
what_day_alt(abbr = TRUE)
what_day_alt(as_integer = TRUE)

# Work with vectors (when as characters):
ds <- c("2020-01-01", "2020-02-29", "2020-12-24", "2020-12-31")
what_day_alt(when = ds)
what_day_alt(when = ds, unit = "month", as_integer = TRUE)
what_day_alt(when = ds, unit = "year", as_integer = TRUE)

 
}
\seealso{
\code{what_day()} for a simpler version (only weekdays); 
\code{what_date()} function to obtain dates; 
\code{what_time()} function to obtain times; 
\code{cur_time()} function to print the current time; 
\code{cur_date()} function to print the current date; 
\code{Sys.time()} function of \strong{base} R.

Other date and time functions: 
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{what_date}()},
\code{\link{what_day}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
