% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun.R
\name{coin}
\alias{coin}
\title{Flip a fair coin (with 2 sides "H" and "T") n times.}
\usage{
coin(n = 1)
}
\arguments{
\item{n}{Number of coin flips.
Default: \code{n = 1}.}
}
\description{
\code{coin} generates a sequence of events that 
represent the results of flipping a fair coin \code{n} times.
}
\details{
The 2 possible outcomes for each flip are 
"H" (for "heads") and "T" (for "tails").
}
\examples{
# Basics: 
coin()
table(coin(n = 100))

# Limits:
coin(2:3)
coin(NA)
coin(0)
coin(1/2) 

}
\seealso{
Other random functions: \code{\link{dice_2}},
  \code{\link{dice}}
}
\concept{random functions}
