% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i2ds_survey}
\alias{i2ds_survey}
\title{i2ds online survey data.}
\format{
A table with 33 participants (rows) and 112 variables (columns).
}
\source{
See online survey at \url{https://ww3.unipark.de/uc/i2ds_survey/}.
}
\usage{
i2ds_survey
}
\description{
\code{i2ds_survey} contains pre-processed data 
from the i2ds online survey.
}
\details{
On 2025-05-15, this data contained 33 participants.

See codebook and print version for questions and coding details.

\strong{Overview} 

The table contains 33 participants (rows) and 112 variables (columns).

Codes for variable name prefixes:

\itemize{

\item \strong{rv}: random variable

\item \strong{c(#)}: choice variable (with # alternatives)

\item \strong{t}: text variable (with any input)

\item \strong{tn}: text variable (with numeric input)

\item \strong{crs}: course-related variable
}

\code{tb} was originally created to practice loops and iterations 
(as a CSV file).
}
\seealso{
Other datasets: 
\code{\link{Bushisms}},
\code{\link{Trumpisms}},
\code{\link{countries}},
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{dt_10}},
\code{\link{exp_num_dt}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{flowery}},
\code{\link{fruits}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{t_1}},
\code{\link{t_2}},
\code{\link{t_3}},
\code{\link{t_4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{table9}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
