% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Is date or time in a leap year?}
\source{
See \url{https://en.wikipedia.org/wiki/Leap_year} for definition.
}
\usage{
is_leap_year(dt)
}
\arguments{
\item{dt}{Date or time (scalar or vector).}
}
\description{
\code{is_leap_year} checks whether a given date or time \code{dt} 
is in a so-called leap year (with a date of February 29).
}
\details{

}
\examples{
is_leap_year(2020)

# from dates:
is_leap_year(Sys.Date())
is_leap_year(as.Date("2022-02-28"))

# from times:
is_leap_year(Sys.time())
is_leap_year(as.POSIXct("2022-10-11 10:11:12"))
is_leap_year(as.POSIXlt("2022-10-11 10:11:12"))

# from non-integers:
is_leap_year(2019.5)

# For vectors:
is_leap_year(2020:2028)

# with dt as strings:
is_leap_year(c("2020", "2021"))
is_leap_year(c("2020-02-29 01:02:03", "2021-02-28 01:02"))
# Note: Invalid date string would yield error
# is_leap_year("2021-02-29")


}
\seealso{
\code{leap_year} function of the \strong{lubridate} package.

Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{what_date}()},
\code{\link{what_day}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
