% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Is date or time in a leap year?}
\source{
See \url{https://en.wikipedia.org/wiki/Leap_year} for definition.
}
\usage{
is_leap_year(dt)
}
\arguments{
\item{dt}{Date or time (scalar or vector).}
}
\description{
\code{is_leap_year} checks whether a given date or time \code{dt} 
is in a so-called leap year (with a date of February 29).
}
\details{
When \code{dt} is not recognized as date or time object(s), 
\code{is_leap_year} aims to interpret it as an integer  
that corresponds to a year (in the 4-digit "%Y" format). 

\code{is_leap_year} solves the task in two ways:  
1. by verifying the numeric definition of a "leap year", and 
2. by trying to use \code{as.Date()} for defining 
a "Date" of Feb-29 in the corresponding year(s).
}
\examples{
# Check:
is_leap_year(2020)    # integer
is_leap_year("2021")  # character

# Dates:
is_leap_year(Sys.Date())
is_leap_year(as.Date("2022-10-11"))

# Times: 
is_leap_year(Sys.time())
is_leap_year(as.POSIXct("2022-10-11 10:11:12"))
is_leap_year(as.POSIXlt("2022-10-11 10:11:12"))

# Note rounding: 
# is_leap_year(2019.5)

# With vectors:
v <- 2020:2028
is_leap_year(v)

}
\seealso{
\code{leap_year} function of the \strong{lubridate} package.

Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{what_date}()},
\code{\link{what_day}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
