% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\docType{data}
\name{l33t_rul35}
\alias{l33t_rul35}
\title{l33t_rul35 provides rules for translating into leet/l33t slang.}
\format{An object of class \code{character} of length 13.}
\usage{
l33t_rul35
}
\description{
\code{l33t_rul35} specifies rules for translating characters 
into leet/l33t slang (as a character vector).
}
\seealso{
Other text functions: \code{\link{transl33t}}
}
\concept{text functions}
\keyword{datasets}
