% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\docType{data}
\name{l33t_rul35}
\alias{l33t_rul35}
\title{l33t_rul35 provides rules for translating text 
into leet/l33t slang.}
\format{
An object of class \code{character} of length 13.
}
\usage{
l33t_rul35
}
\description{
\code{l33t_rul35} specifies rules for translating characters 
into other characters (typically symbols) to mimic 
leet/l33t slang (as a named character vector).
}
\details{
Old (i.e., to be replaced) characters are 
\code{paste(names(l33t_rul35), collapse = "")}.

New (i.e., replaced) characters are 
\code{paste(l33t_rul35, collapse = "")}.

See \url{https://en.wikipedia.org/wiki/Leet} for details.
}
\seealso{
\code{\link{transl33t}} for a corresponding function.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{cclass}},
\code{\link{count_chars_words}()},
\code{\link{count_chars}()},
\code{\link{count_words}()},
\code{\link{invert_rules}()},
\code{\link{map_text_chars}()},
\code{\link{map_text_coord}()},
\code{\link{map_text_regex}()},
\code{\link{metachar}},
\code{\link{read_ascii}()},
\code{\link{text_to_chars}()},
\code{\link{text_to_sentences}()},
\code{\link{text_to_words}()},
\code{\link{transl33t}()},
\code{\link{words_to_text}()}
}
\concept{text objects and functions}
\keyword{datasets}
