% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\docType{data}
\name{metachar}
\alias{metachar}
\title{metachar provides metacharacters (as a character vector).}
\format{
An object of class \code{character} of length 12.
}
\usage{
metachar
}
\description{
\code{metachar} provides the metacharacters of extended regular expressions 
(as a character vector).
}
\details{
\code{metachar} allows illustrating the notion of 
meta-characters in regular expressions 
(and provides corresponding exemplars). 

See \code{?base::regex} for details on regular expressions 
and \code{?"'"} for a list of character constants/quotes in R.
}
\examples{
metachar
length(metachar)  # 12
nchar(paste0(metachar, collapse = ""))  # 12

}
\seealso{
\code{\link{cclass}} for a vector of character classes.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{cclass}},
\code{\link{count_chars_words}()},
\code{\link{count_chars}()},
\code{\link{count_words}()},
\code{\link{l33t_rul35}},
\code{\link{map_text_chars}()},
\code{\link{map_text_coord}()},
\code{\link{map_text_regex}()},
\code{\link{read_ascii}()},
\code{\link{text_to_chars}()},
\code{\link{text_to_sentences}()},
\code{\link{text_to_words}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
\keyword{datasets}
