% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{data}
\name{pal_ds4psy}
\alias{pal_ds4psy}
\title{ds4psy default color palette.}
\format{An object of class \code{data.frame} with 1 rows and 11 columns.}
\usage{
pal_ds4psy
}
\description{
\code{pal_ds4psy} provides a dedicated color palette.
}
\details{
By default, \code{pal_ds4psy} is initialized to 
\code{pal_unikn} of the \bold{unikn} package.
}
\seealso{
Other plot functions: \code{\link{pal_n_sq}},
  \code{\link{theme_ds4psy}}
}
\concept{plot functions}
\keyword{datasets}
