% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posPsy_long}
\alias{posPsy_long}
\title{Positive Psychology: AHI_CESD corrected data (in long format).}
\format{A tibble with 990 cases (rows) and 50 variables (columns).
 
\strong{Codebook}  

See \code{\link{posPsy_AHI_CESD}}.}
\source{
\strong{Articles}

Radloff, L. S. (1977). 
The CES-D scale: A self report depression scale for research in the general population. 
\emph{Applied Psychological Measurement}, \emph{1}, 385–401. 
doi: \code{10.1177/014662167700100306} 

Woodworth, R. J., O’Brien‐Malone, A., Diamond, M. R., & Schüz, B. (2017). 
Web‐based positive psychology interventions: A reexamination of effectiveness. 
\emph{Journal of Clinical Psychology}, \emph{73}(3), 218–232. 
doi: \code{10.1002/jclp.22328} 

Woodworth, R. J., O’Brien-Malone, A., Diamond, M. R. and Schüz, B. (2018). 
Data from, ‘Web-based positive psychology interventions: A reexamination of effectiveness’. 
\emph{Journal of Open Psychology Data}, \emph{6}(1). 
doi: \code{10.5334/jopd.35} 

\strong{Data}

See \url{https://openpsychologydata.metajnl.com/articles/10.5334/jopd.35/} for details.

Original dataset available at \url{https://doi.org/10.6084/m9.figshare.1577563.v1}.
}
\usage{
posPsy_long
}
\description{
This dataset contains answers to the 24 items of the 
Authentic Happiness Inventory (AHI) and answers to the 
20 items of the Center for Epidemiological Studies Depression (CES-D) scale 
(see Radloff, 1977) for multiple (1 to 6) measurement occasions.
}
\seealso{
\code{\link{posPsy_AHI_CESD}} for source of this file, 
\code{\link{posPsy_wide}} for a version of this file (in wide format).

Other datasets: \code{\link{falsePosPsy_all}},
  \code{\link{posPsy_AHI_CESD}},
  \code{\link{posPsy_p_info}}, \code{\link{posPsy_wide}}
}
\concept{datasets}
\keyword{datasets}
