% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posPsy_wide}
\alias{posPsy_wide}
\title{Positive Psychology: All corrected data (in wide format).}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 295 rows and 294 columns.}
\source{
\strong{Articles}

Radloff, L. S. (1977). 
The CES-D scale: A self report depression scale for research in the general population. 
\emph{Applied Psychological Measurement}, \emph{1}, 385–401. 
doi: \code{10.1177/014662167700100306} 

Woodworth, R. J., O’Brien‐Malone, A., Diamond, M. R., & Schüz, B. (2017). 
Web‐based positive psychology interventions: A reexamination of effectiveness. 
\emph{Journal of Clinical Psychology}, \emph{73}(3), 218–232. 
doi: \code{10.1002/jclp.22328} 

Woodworth, R. J., O’Brien-Malone, A., Diamond, M. R. and Schüz, B. (2018). 
Data from, ‘Web-based positive psychology interventions: A reexamination of effectiveness’. 
\emph{Journal of Open Psychology Data}, \emph{6}(1). 
doi: \code{10.5334/jopd.35} 

\strong{Data}

See \url{https://openpsychologydata.metajnl.com/articles/10.5334/jopd.35/} for details.

Original dataset available at \url{https://doi.org/10.6084/m9.figshare.1577563.v1}.
}
\usage{
posPsy_wide
}
\description{
Positive Psychology: All corrected data (in wide format).
}
\seealso{
\code{\link{posPsy_AHI_CESD}} for the source of this file, 
\code{\link{posPsy_long}} for a version of this file (in long format).

Other datasets: \code{\link{falsePosPsy_all}},
  \code{\link{posPsy_AHI_CESD}}, \code{\link{posPsy_long}},
  \code{\link{posPsy_p_info}}
}
\concept{datasets}
\keyword{datasets}
