% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{read_ascii}
\alias{read_ascii}
\title{read_ascii parses text (from a file) into a tibble.}
\usage{
read_ascii(file = "txt/ascii.txt", flip_y = FALSE)
}
\arguments{
\item{file}{The text file to read (or its path).
If the text file is stored in a sub-directory, 
enter its path and name here (without any leading or 
trailing "." or "/"). 
Default: \code{file = "txt/ascii.txt"}.}

\item{flip_y}{Boolean: Should y-coordinates be flipped, 
so that the lowest line in the text file becomes \code{y = 1}, 
and the top line in the text file becomes \code{y = n_lines}? 
Default: \code{flip_y = FALSE}.}
}
\description{
\code{read_ascii} parses text (from a file) into 
a tibble that contains a row for each character. 
This tibble contains 3 variables: 
The character's \code{x}- and \code{y}-coordinates (from top to bottom)  
and a variable \code{char} for the character at this coordinate.
}
\details{
The \bold{here} package is used to determine 
the (absolute) file path.
}
\examples{
read_ascii("txt/ascii.txt")  # requires txt file
read_ascii("txt/ascii.txt", flip_y = TRUE)

}
\seealso{
corresponding plot function

Other text functions: \code{\link{l33t_rul35}},
  \code{\link{transl33t}}
}
\concept{text functions}
