% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_fun.R
\name{text_to_sentences}
\alias{text_to_sentences}
\title{text_to_sentences turns a string of text \code{x} 
(consisting of one or more character strings) 
into a vector of its sentences.}
\usage{
text_to_sentences(x)
}
\arguments{
\item{x}{A string of text (required).}
}
\description{
text_to_sentences removes all (standard) punctuation marks and empty spaces 
and returns a vector of all remaining character sequences  
(as the sentences).
}
\examples{
# Default: 
x <- c("Hello!", "This is a 1st sentence.", "Is this a question?", "The end.")
text_to_sentences(x)

}
\seealso{
\code{\link{text_to_words}} for turning text into a vector of words.

Other text objects and functions: 
\code{\link{Umlaut}},
\code{\link{capitalize}()},
\code{\link{caseflip}()},
\code{\link{count_char}()},
\code{\link{count_word}()},
\code{\link{l33t_rul35}},
\code{\link{metachar}},
\code{\link{read_ascii}()},
\code{\link{text_to_words}()},
\code{\link{transl33t}()}
}
\concept{text objects and functions}
