% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fun.R
\name{theme_ds4psy}
\alias{theme_ds4psy}
\title{ds4psy default plot theme (using ggplot2 and unikn).}
\usage{
theme_ds4psy(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  col_title = grey(0, 1),
  col_txt_1 = grey(0.1, 1),
  col_txt_2 = grey(0.2, 1),
  col_txt_3 = grey(0.1, 1),
  col_bg = "transparent"
)
}
\arguments{
\item{base_size}{Base font size (optional, numeric). 
Default: \code{base_size = 11}.}

\item{base_family}{Base font family (optional, character). 
Default: \code{base_family = ""}.}

\item{base_line_size}{Base line size (optional, numeric). 
Default: \code{base_line_size = base_size/20}.}

\item{base_rect_size}{Base rectangle size (optional, numeric). 
Default: \code{base_rect_size = base_size/20}.}

\item{col_title}{Color of plot title (text) elements (optional).
Default: \code{col_title = grey(.0, 1)} (i.e., "black").}

\item{col_txt_1}{Color of text elements of headings and axis labels (optional).
Default: \code{col_title = grey(.1, 1)}.}

\item{col_txt_2}{Color of text elements of legend and axis text/ticks (optional).
Default: \code{col_title = grey(.2, 1)}.}

\item{col_txt_3}{Color of text elements of facet strips (optional).
Default: \code{col_title = grey(.1, 1)}.}

\item{col_bg}{Color of plot background (optional).
Default: \code{col_bg = "transparent"}.}
}
\description{
\code{theme_ds4psy} provides a basic \bold{ds4psy} theme 
to use in \bold{ggplot2} commands.
}
\details{
The theme is lightweight and no-nonsense, but somewhat 
opinionated (e.g., in using mostly grey scales to 
allow emphasizing data points with color accents). 

Whereas the colors of text elements and the plot background can be specified, 
the panel background is assumed to be white.
}
\examples{

\donttest{
  # Plotting iris dataset (using ggplot2 and theme_ds4psy):
  
  library("ggplot2")  # theme_ds4psy requires loading ggplot2 
  
  ggplot(datasets::iris) +
    geom_jitter(aes(x = Petal.Length, y = Petal.Width, color = Species), size = 3, alpha = 2/3) +
    labs(title = "Iris species",
         caption = "Data from datasets::iris") +
    theme_ds4psy(col_title = "black", base_size = 11)
}

}
\seealso{
\code{unikn::theme_unikn} for the source of the current theme.

Other plot functions: 
\code{\link{plot_fn}()},
\code{\link{plot_fun}()},
\code{\link{plot_n}()},
\code{\link{plot_text}()},
\code{\link{plot_tiles}()}
}
\concept{plot functions}
