% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trumpisms}
\alias{Trumpisms}
\title{Data: Trumpisms.}
\format{
A vector of type \code{character}  
with \code{length(Trumpisms) = 168} 
(on 2021-01-28).
}
\source{
Data originally based on a collection of \emph{Donald Trump's 20 most frequently used words} on \url{https://www.yourdictionary.com}  
and expanded by interviews, public speeches, and Twitter tweets from \code{https://twitter.com/realDonaldTrump}.
}
\usage{
Trumpisms
}
\description{
\code{Trumpisms} contains frequent words and characteristic phrases 
by U.S. president Donald J. Trump (the 45th president of the United States, 
in office from January 20, 2017, to January 20, 2021).
}
\seealso{
Other datasets: 
\code{\link{Bushisms}},
\code{\link{countries}},
\code{\link{data_1}},
\code{\link{data_2}},
\code{\link{data_t1_de}},
\code{\link{data_t1_tab}},
\code{\link{data_t1}},
\code{\link{data_t2}},
\code{\link{data_t3}},
\code{\link{data_t4}},
\code{\link{dt_10}},
\code{\link{exp_num_dt}},
\code{\link{exp_wide}},
\code{\link{falsePosPsy_all}},
\code{\link{fame}},
\code{\link{flowery}},
\code{\link{fruits}},
\code{\link{outliers}},
\code{\link{pi_100k}},
\code{\link{posPsy_AHI_CESD}},
\code{\link{posPsy_long}},
\code{\link{posPsy_p_info}},
\code{\link{posPsy_wide}},
\code{\link{t3}},
\code{\link{t4}},
\code{\link{t_1}},
\code{\link{t_2}},
\code{\link{t_3}},
\code{\link{t_4}},
\code{\link{table6}},
\code{\link{table7}},
\code{\link{table8}},
\code{\link{table9}},
\code{\link{tb}}
}
\concept{datasets}
\keyword{datasets}
