% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{what_age}
\alias{what_age}
\title{What age does some date(s) have today (in full years)?}
\usage{
what_age(dt)
}
\arguments{
\item{dt}{Date or time (scalar or vector). 
Numbers or strings with dates are parsed into 
4-digit numbers denoting the year}
}
\description{
\code{what_age} provides the number of completed years 
for a (vector of type) "Date".
}
\details{
\code{dt} is assumed to be of class "Date" 
and co-erced into "Date" when of class "POSIXt".
}
\examples{
y_100 <- Sys.Date() - (100 * 365.25) + -1:1
what_age(y_100)

}
\seealso{
Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{is_leap_year}()},
\code{\link{what_date}()},
\code{\link{what_day}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
