% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_fun.R
\name{what_age}
\alias{what_age}
\title{What is the age (or some date difference) in human units?}
\usage{
what_age(from_date, to_date = Sys.Date(), units = "y")
}
\arguments{
\item{from_date}{From date (required, scalar or vector, as "Date"). 
Date of birth (DOB), assumed to be of class "Date", 
and coerced into "Date" when of class "POSIXt".}

\item{to_date}{To date (optional, scalar or vector, as "Date"). 
Default: \code{to_date = Sys.Date()}. 
Maximum date/date of death (DOD), assumed to be of class "Date", 
and coerced into "Date" when of class "POSIXt".}

\item{units}{Units used to represent output (as "character").
Units represent human time periods, rather than 
chronological time differences. 
Default: \code{units = "y"} for "years". 
Available options include:  

\enumerate{

  \item \code{units = "y"}: completed years (default)
  
  \item \code{units = "ym"}: completed years, months
  
  \item \code{units = "ymd"}: completed years, months, days
  
  }}
}
\description{
\code{what_age} provides the difference between two dates 
(i.e., from some \code{from_date} to some \code{to_date}) 
in human measurement units (periods).
}
\details{
If not specified explicitly, \code{to_date} is set to 
today's date (i.e., \code{Sys.Date()}).

If the lengths of \code{from_date} and \code{to_date} differ, 
the arguments of \code{to_date} are recycled or 
truncated to the length of \code{from_date}.
}
\examples{
y_100 <- Sys.Date() - (100 * 365.25) + -1:1
what_age(y_100)

# with "to_date" argument: 
y_050 <- Sys.Date() - (50 * 365.25) + -1:1 
what_age(y_100, y_050)

# robustness:
days_this_year <- 365 + is_leap_year(Sys.Date())
what_age(Sys.time() - (10 * (60 * 60 * 24) * days_this_year)) # for POSIXt times
what_age("90-07-11", to_date = "10-07-10")                    # for strings
what_age(19900711, to_date = 20100710)                        # for numbers

# recycling "to_date" to length of "from_date":
y_050_2 <- Sys.Date() - (50 * 365.25)
what_age(y_100, y_050_2)

# Using 'fame' data:
dob <- as.Date(fame$DOB, format = "\%B \%d, \%Y")
dod <- as.Date(fame$DOD, format = "\%B \%d, \%Y")
what_age(dob, dod)  # Note: Deceased people do not age further.

}
\seealso{
Other date and time functions: 
\code{\link{change_time}()},
\code{\link{change_tz}()},
\code{\link{cur_date}()},
\code{\link{cur_time}()},
\code{\link{is_leap_year}()},
\code{\link{what_date}()},
\code{\link{what_month}()},
\code{\link{what_time}()},
\code{\link{what_wday}()},
\code{\link{what_week}()},
\code{\link{what_year}()}
}
\concept{date and time functions}
